Restructure the project folders; includes name change for FileHandler component
