+++ src/components/ProgressBar/ProgressBar.styles.js
